:recycle: Refactor: CustomAlert 하나로 통합
